refactor: UNT-T13160 Update the skia ThemeSwitch component - Refactor Skia Theme Switch Component
