+++ example/src/modules/ThemeSwitch/ThemeSwitchScreen.tsx
@@ -1,11 +1,10 @@
 import {useNavigation} from '@react-navigation/native';
 import React from 'react';
 import {View} from 'react-native';
-import {SkiaThemeSwitch} from 'react-native-skia-catalog';
+import {ThemeSwitch} from 'react-native-skia-catalog';
 import {CustomHeader} from '../../components';
 import {Strings} from '../../constants';
 import {NavProps} from '../../navigation/types';
-import {Colors} from '../../theme';
 import styles from './styles/styles';
 
 const ThemeSwitchScreen = () => {
@@ -19,7 +18,7 @@ const ThemeSwitchScreen = () => {
         onBackPress={() => navigation.goBack()}
       />
       <View style={styles.container}>
-        <SkiaThemeSwitch />
+        <ThemeSwitch />
       </View>
     </View>
   );
